update foeach for set and map
